Finished

+++ tutorial-5/src/main/java/csui/advpro2021/tais/service/MahasiswaServiceImpl.java
@@ -35,6 +35,10 @@ public class MahasiswaServiceImpl implements MahasiswaService {
 
     @Override
     public void deleteMahasiswaByNPM(String npm) {
-        mahasiswaRepository.deleteById(npm);
+        Mahasiswa mahasiswa = mahasiswaRepository.findByNpm(npm);
+        if (mahasiswa.getMataKuliah() == null) {
+            mahasiswaRepository.deleteById(npm);
+        }
+
     }
 }
